inventory: move to own dir; add vars
