+++ meson.build
@@ -96,7 +96,8 @@ dll_sources = [
   'src/game/gun/gun_rifle.c',
   'src/game/hwr.c',
   'src/game/input.c',
-  'src/game/inventory.c',
+  'src/game/inventory/common.c',
+  'src/game/inventory/vars.c',
   'src/game/items.c',
   'src/game/lara/lara_cheat.c',
   'src/game/lara/lara_col.c',
