+++ src/game/gameflow.c
@@ -5,7 +5,7 @@
 #include "game/demo.h"
 #include "game/game.h"
 #include "game/gun/gun.h"
-#include "game/inventory.h"
+#include "game/inventory/common.h"
 #include "game/objects/names.h"
 #include "game/overlay.h"
 #include "gameflow/gameflow_new.h"
