+++ src/game/inventory/common.c
@@ -1,7 +1,8 @@
-#include "game/inventory.h"
+#include "game/inventory/common.h"
 
 #include "game/demo.h"
 #include "game/input.h"
+#include "game/inventory/vars.h"
 #include "game/items.h"
 #include "game/lara/lara_control.h"
 #include "game/matrix.h"
