+++ src/game/inventory/common.h
