+++ src/game/inventory/vars.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "global/types.h"
+
+extern int32_t g_Inv_OptionObjectsCount;
+extern INVENTORY_ITEM *g_Inv_OptionList[];
