+++ src/game/lara/lara_cheat.c
@@ -5,7 +5,7 @@
 #include "game/console.h"
 #include "game/creature.h"
 #include "game/game_string.h"
-#include "game/inventory.h"
+#include "game/inventory/common.h"
 #include "game/items.h"
 #include "game/lara/lara_control.h"
 #include "game/math.h"
