+++ src/game/lara/lara_control.c
@@ -2,7 +2,7 @@
 
 #include "game/creature.h"
 #include "game/gun/gun.h"
-#include "game/inventory.h"
+#include "game/inventory/common.h"
 #include "game/items.h"
 #include "game/lara/lara_cheat.h"
 #include "game/lara/lara_look.h"
