+++ src/inject_exec.c
@@ -18,7 +18,7 @@
 #include "game/gun/gun_rifle.h"
 #include "game/hwr.h"
 #include "game/input.h"
-#include "game/inventory.h"
+#include "game/inventory/common.h"
 #include "game/items.h"
 #include "game/lara/lara_cheat.h"
 #include "game/lara/lara_col.h"
